chore: rename dev/util to dev/internal/util
