+++ dev/docs.go
@@ -6,7 +6,7 @@ import (
 	"golang.org/x/sync/errgroup"
 
 	"github.com/dagger/dagger/dev/internal/dagger"
-	"github.com/dagger/dagger/dev/util"
+	"github.com/dagger/dagger/dev/internal/util"
 )
 
 type Docs struct {
