+++ dev/sdk_php.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/dagger/dagger/dev/internal/dagger"
-	"github.com/dagger/dagger/dev/util"
+	"github.com/dagger/dagger/dev/internal/util"
 )
 
 const (
