+++ dev/sdk_typescript.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/dagger/dagger/dev/internal/build"
 	"github.com/dagger/dagger/dev/internal/dagger"
-	"github.com/dagger/dagger/dev/util"
+	"github.com/dagger/dagger/dev/internal/util"
 )
 
 // TODO: use dev module (this is just the mage port)
